[ADD] Use IPC signaling to stop directory during E2E tests on Windows

+++ pandora-tests/test/utils/server.ts
@@ -1,5 +1,5 @@
 import { Assert, AssertNotNullable, EnvStringify } from './utils';
-import { ChildProcessWithoutNullStreams, spawn } from 'child_process';
+import { spawn, fork, type ChildProcess } from 'child_process';
 import { PNPM_EXECUTABLE, TEST_COVERAGE_TEMP, TEST_DIRECTORY_PORT, TEST_PROJECT_PANDORA_DIR, TEST_SERVER_DIRECTORY_ENTRYPOINT, TEST_SERVER_DIRECTORY_TEST_DIR, TEST_TEMP } from '../_setup/config';
 import { test } from '@playwright/test';
 import path from 'path';
@@ -40,7 +40,7 @@ export interface TestStartDirectoryOptions {
 }
 
 type ServerInstanceData = {
-	process: ChildProcessWithoutNullStreams;
+	process: ChildProcess;
 	keepActive: boolean;
 	stdout: string;
 	stderr: string;
@@ -78,7 +78,12 @@ export function TestStopDirectory(): Promise<void> {
 			resolve();
 		});
 		server.process.off('exit', HandlePrematureExit);
-		server.process.kill('SIGINT');
+		// On Windows soft-kill doesn't really work; we need to use IPC
+		if (process.platform === 'win32') {
+			server.process.send('STOP');
+		} else {
+			server.process.kill('SIGINT');
+		}
 	});
 }
 
@@ -99,26 +104,41 @@ export function InternalSetupTestingEnvDirectory(): void {
 export function TestStartDirectory(options: Partial<TestStartDirectoryOptions> = {}): Promise<void> {
 	Assert(DirectoryServer == null);
 
-	const directoryProcess = spawn(PNPM_EXECUTABLE, [
-		'exec',
-		'nyc',
-		'--silent',
-		'--no-clean', // Clean is handled by global setup
-		'--cwd', TEST_PROJECT_PANDORA_DIR, // This is "working directory" only for istanbul, not for rest of command
-		'--temp-dir', TEST_COVERAGE_TEMP, // This needs to be an absolute path
-		'node',
-		'--enable-source-maps',
-		TEST_SERVER_DIRECTORY_ENTRYPOINT,
-	], {
-		cwd: TEST_TEMP,
-		env: EnvStringify({
-			...process.env,
-			...DIRECTORY_ENV_DEFAULTS,
-			...options.configOverrides,
-		}),
-		shell: true,
-		stdio: 'pipe',
-	});
+	let directoryProcess: ChildProcess;
+	// On Windows we need IPC capabilities, so use fork
+	// Using this we loose on nyc (coverage collection), but there is no way around this
+	if (process.platform === 'win32') {
+		directoryProcess = fork(TEST_SERVER_DIRECTORY_ENTRYPOINT, {
+			cwd: TEST_PROJECT_PANDORA_DIR,
+			env: EnvStringify({
+				...process.env,
+				...DIRECTORY_ENV_DEFAULTS,
+				...options.configOverrides,
+			}),
+			stdio: 'pipe',
+		});
+	} else {
+		directoryProcess = spawn(PNPM_EXECUTABLE, [
+			'exec',
+			'nyc',
+			'--silent',
+			'--no-clean', // Clean is handled by global setup
+			'--cwd', TEST_PROJECT_PANDORA_DIR, // This is "working directory" only for istanbul, not for rest of command
+			'--temp-dir', TEST_COVERAGE_TEMP, // This needs to be an absolute path
+			'node',
+			'--enable-source-maps',
+			TEST_SERVER_DIRECTORY_ENTRYPOINT,
+		], {
+			cwd: TEST_TEMP,
+			env: EnvStringify({
+				...process.env,
+				...DIRECTORY_ENV_DEFAULTS,
+				...options.configOverrides,
+			}),
+			shell: true,
+			stdio: 'pipe',
+		});
+	}
 
 	const instanceData: ServerInstanceData = {
 		process: directoryProcess,
@@ -129,10 +149,10 @@ export function TestStartDirectory(options: Partial<TestStartDirectoryOptions> =
 
 	DirectoryServer = instanceData;
 
-	directoryProcess.stdout.on('data', function (data: string | Buffer) {
+	directoryProcess.stdout?.on('data', function (data: string | Buffer) {
 		instanceData.stdout += data.toString();
 	});
-	directoryProcess.stderr.on('data', function (data: string | Buffer) {
+	directoryProcess.stderr?.on('data', function (data: string | Buffer) {
 		instanceData.stderr += data.toString();
 	});
 	directoryProcess.on('exit', HandlePrematureExit);
